requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
google-generativeai==0.3.2
faiss-cpu==1.13.2 
numpy==1.26.4
beautifulsoup4==4.12.3
requests==2.31.0
pydantic==2.6.1
python-multipart==0.0.9
aiofiles==23.2.1
